+++ devices/SparkfunLCD/SparkfunLCD.nfproj
@@ -15,7 +15,7 @@
     <RootNamespace>Iot.Device.SparkfunLcd</RootNamespace>
     <AssemblyName>Iot.Device.SparkfunLcd</AssemblyName>
     <TargetFrameworkVersion>v1.0</TargetFrameworkVersion>
-    <DocumentationFile>bin\$(Configuration)\Iot.Device.SparkfunLCD.xml</DocumentationFile>
+    <DocumentationFile>bin\$(Configuration)\Iot.Device.SparkfunLcd.xml</DocumentationFile>
     <LangVersion>9.0</LangVersion>
   </PropertyGroup>
   <PropertyGroup>
@@ -41,6 +41,8 @@
       <ExcludeFromStyleCop>true</ExcludeFromStyleCop>
       <Link>ICharacterLcd.cs</Link>
     </Compile>
+    <Compile Include="DisplaySizeEnum.cs" />
+    <Compile Include="OpenLcdCommandEnum.cs" />
     <Compile Include="SparkfunLcd.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <None Include="*.md" />
